Moving propagation api into propagation/__init__.py
