+++ opentelemetry-api/src/opentelemetry/propagation/__init__.py
@@ -12,24 +12,159 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import abc
 import typing
 
-import opentelemetry.context.propagation.httptextformat as httptextformat
 import opentelemetry.trace as trace
 from opentelemetry.context import Context
-from opentelemetry.context.propagation import (
-    ContextT,
-    DefaultExtractor,
-    DefaultInjector,
-    Getter,
-    Setter,
-)
+
+ContextT = typing.TypeVar("ContextT")
+
+Setter = typing.Callable[[ContextT, str, str], None]
+Getter = typing.Callable[[ContextT, str], typing.List[str]]
+
+
+class Extractor(abc.ABC):
+    """API for propagation of span context via headers.
+
+    TODO: update docs to reflect split into extractor/injector
+
+    This class provides an interface that enables extracting and injecting
+    span context into headers of HTTP requests. HTTP frameworks and clients
+    can integrate with HTTPTextFormat by providing the object containing the
+    headers, and a getter and setter function for the extraction and
+    injection of values, respectively.
+
+    Example::
+
+        import flask
+        import requests
+        from opentelemetry.context.propagation import Extractor
+
+        PROPAGATOR = HTTPTextFormat()
+
+
+
+        def get_header_from_flask_request(request, key):
+            return request.headers.get_all(key)
+
+        def set_header_into_requests_request(request: requests.Request,
+                                             key: str, value: str):
+            request.headers[key] = value
+
+        def example_route():
+            span_context = PROPAGATOR.extract(
+                get_header_from_flask_request,
+                flask.request
+            )
+            request_to_downstream = requests.Request(
+                "GET", "http://httpbin.org/get"
+            )
+            PROPAGATOR.inject(
+                span_context,
+                set_header_into_requests_request,
+                request_to_downstream
+            )
+            session = requests.Session()
+            session.send(request_to_downstream.prepare())
+
+
+    .. _Propagation API Specification:
+       https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/api-propagators.md
+    """
+
+    @classmethod
+    @abc.abstractmethod
+    def extract(
+        cls,
+        carrier: ContextT,
+        context: typing.Optional[Context] = None,
+        get_from_carrier: typing.Optional[Getter[ContextT]] = None,
+    ) -> Context:
+        """Create a Context from values in the carrier.
+
+        The extract function should retrieve values from the carrier
+        object using get_from_carrier, use values to populate a
+        Context value and return it.
+
+        Args:
+            carrier: and object which contains values that are
+                used to construct a Context. This object
+                must be paired with an appropriate get_from_carrier
+                which understands how to extract a value from it.
+            context: The Context to read values from.
+            get_from_carrier: a function that can retrieve zero
+                or more values from the carrier. In the case that
+                the value does not exist, return an empty list.
+        Returns:
+            A Context with configuration found in the carrier.
+        """
+
+
+class Injector(abc.ABC):
+    @classmethod
+    @abc.abstractmethod
+    def inject(
+        cls,
+        carrier: ContextT,
+        context: typing.Optional[Context] = None,
+        set_in_carrier: typing.Optional[Setter[ContextT]] = None,
+    ) -> None:
+        """Inject values from a Context into a carrier.
+
+        inject enables the propagation of values into HTTP clients or
+        other objects which perform an HTTP request. Implementations
+        should use the set_in_carrier method to set values on the
+        carrier.
+
+        Args:
+            carrier: An object that a place to define HTTP headers.
+                Should be paired with set_in_carrier, which should
+                know how to set header values on the carrier.
+            context: The Context to read values from.
+            set_in_carrier: A setter function that can set values
+                on the carrier.
+        """
+
+
+class DefaultExtractor(Extractor):
+    """The default Extractor that is used when no Extractor implementation is configured.
+
+    All operations are no-ops.
+    """
+
+    @classmethod
+    def extract(
+        cls,
+        carrier: ContextT,
+        context: typing.Optional[Context] = None,
+        get_from_carrier: typing.Optional[Getter[ContextT]] = None,
+    ) -> Context:
+        if context:
+            return context
+        return Context.current()
+
+
+class DefaultInjector(Injector):
+    """The default Injector that is used when no Injector implementation is configured.
+
+    All operations are no-ops.
+    """
+
+    @classmethod
+    def inject(
+        cls,
+        carrier: ContextT,
+        context: typing.Optional[Context] = None,
+        set_in_carrier: typing.Optional[Setter[ContextT]] = None,
+    ) -> None:
+        return None
 
 
 def extract(
     carrier: ContextT,
     context: typing.Optional[Context] = None,
-    extractors: typing.Optional[typing.List[httptextformat.Extractor]] = None,
+    extractors: typing.Optional[typing.List[Extractor]] = None,
     get_from_carrier: typing.Optional[Getter[ContextT]] = None,
 ) -> typing.Optional[Context]:
     """Load the parent SpanContext from values in the carrier.
@@ -67,7 +202,7 @@ def extract(
 
 def inject(
     carrier: ContextT,
-    injectors: typing.Optional[typing.List[httptextformat.Injector]] = None,
+    injectors: typing.Optional[typing.List[Injector]] = None,
     context: typing.Optional[Context] = None,
 ) -> None:
     """Inject values from the current context into the carrier.
@@ -102,9 +237,7 @@ _HTTP_TEXT_EXTRACTORS = [
 ]  # typing.List[httptextformat.Extractor]
 
 
-def set_http_extractors(
-    extractor_list: typing.List[httptextformat.Extractor],
-) -> None:
+def set_http_extractors(extractor_list: typing.List[Extractor],) -> None:
     """
     To update the global extractor, the Propagation API provides a
     function which takes an extractor.
@@ -113,9 +246,7 @@ def set_http_extractors(
     _HTTP_TEXT_EXTRACTORS = extractor_list  # type: ignore
 
 
-def set_http_injectors(
-    injector_list: typing.List[httptextformat.Injector],
-) -> None:
+def set_http_injectors(injector_list: typing.List[Injector],) -> None:
     """
     To update the global injector, the Propagation API provides a
     function which takes an injector.
@@ -124,7 +255,7 @@ def set_http_injectors(
     _HTTP_TEXT_INJECTORS = injector_list  # type: ignore
 
 
-def get_http_extractors() -> typing.List[httptextformat.Extractor]:
+def get_http_extractors() -> typing.List[Extractor]:
     """
     To access the global extractor, the Propagation API provides
     a function which returns an extractor.
@@ -132,9 +263,26 @@ def get_http_extractors() -> typing.List[httptextformat.Extractor]:
     return _HTTP_TEXT_EXTRACTORS  # type: ignore
 
 
-def get_http_injectors() -> typing.List[httptextformat.Injector]:
+def get_http_injectors() -> typing.List[Injector]:
     """
     To access the global injector, the Propagation API provides a
     function which returns an injector.
     """
     return _HTTP_TEXT_INJECTORS  # type: ignore
+
+
+def get_as_list(
+    dict_object: typing.Dict[str, str], key: str
+) -> typing.List[str]:
+    value = dict_object.get(key)
+    if value is None:
+        return []
+    if isinstance(value, list):
+        return value
+    return [value]
+
+
+def set_in_dict(
+    dict_object: typing.Dict[str, str], key: str, value: str
+) -> None:
+    dict_object[key] = value
