+++ opentelemetry-sdk/src/opentelemetry/sdk/context/propagation/tracecontexthttptextformat.py
@@ -17,7 +17,7 @@ import typing
 
 import opentelemetry.trace as trace
 from opentelemetry.context import Context
-from opentelemetry.context.propagation import (
+from opentelemetry.propagation import (
     Extractor,
     Getter,
     Injector,
